requirements.txt
ollama==0.4.7
gradio==5.23.3
requests==2.32.3
pyyaml==6.0.2
tenacity==9.0.0
